api/requirements.txt
anyio==3.6.2
    # via starlette
attrs==22.2.0
    # via pytest
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
colorama==0.4.6
    # via
    #   loguru
    #   pytest
exceptiongroup==1.1.0
    # via pytest
fastapi==0.92.0
    # via
    #   -r requirements.in
    #   fastapi-paginate
fastapi-paginate==0.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
loguru==0.6.0
    # via -r requirements.in
mangum==0.17.0
    # via -r requirements.in
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
pydantic==1.10.5
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-paginate
pytest==7.2.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.25.0
    # via fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.5.0
    # via
    #   mangum
    #   pydantic
    #   starlette
urllib3==1.26.14
    # via requests
win32-setctime==1.1.0
    # via loguru
httpx==0.23.3
uvicorn==0.20.0